+++ CHANGELOG.rst
@@ -4,7 +4,8 @@ Changelog for package grepros
 
 0.3.4 (2021-12-13)
 -------------------
-* speed up YAML formatting
+* speed up SQLite output (~4x)
+* speed up YAML formatting (~2x)
 * fix no engine name in console texts for Postgres output
 
 0.3.3 (2021-12-13)
@@ -17,7 +18,7 @@ Changelog for package grepros
 * raise Postgres default commit-interval from 100 to 1000
 * add --write-option commit-interval support to SQLite output
 * add --write-option subtypes=array|all, for Postgres/SQLite output
-* speed up SQLite output
+* speed up SQLite output (~2x)
 * refactor Postgres/SQLite sinks onto common base class
 
 0.3.1 (2021-12-06)